fix: EulerAnglesToAxisVector was using sin/cos instead of sinf/cosf

+++ UDT_DLL/src/math.cpp
@@ -78,11 +78,11 @@ namespace Float3
 	void EulerAnglesToAxisVector(f32* result, const f32* angles)
 	{
 		const f32 angle1 = angles[YAW] * (UDT_PI * 2 / 360);
-		const f32 sy = sin(angle1);
-		const f32 cy = cos(angle1);
+		const f32 sy = sinf(angle1);
+		const f32 cy = cosf(angle1);
 		const f32 angle2 = angles[PITCH] * (UDT_PI * 2 / 360);
-		const f32 sp = sin(angle2);
-		const f32 cp = cos(angle2);
+		const f32 sp = sinf(angle2);
+		const f32 cp = cosf(angle2);
 
 		result[0] = cp*cy;
 		result[1] = cp*sy;
